Fix circular dependency in testing suite
