+++ tests/functional/test_appointments.py
@@ -25,10 +25,10 @@ def test_apt_new(client, apt_finput):
     assert response.status_code == 200
 
 
-def test_apt_new_post(client, apt_finput, token):
+def test_apt_new_post(client, apt_finput, token_fixture):
     """Test apt_new endpoint."""
     ppt_id = apt_finput["inputId"]
-    ppt = api.get_participant(ppt_id, token=token)
+    ppt = api.get_participant(ppt_id, token=token_fixture)
     apt_ids = list(ppt.appointments.records.keys())
     last_apt_id = apt_ids[-1].split(":")[1]
     next_apt_id = str(int(last_apt_id) + 1)
@@ -36,7 +36,7 @@ def test_apt_new_post(client, apt_finput, token):
     url = f"/appointments/appointment_new?ppt_id={ppt_id}"
     response = client.post(url, data=apt_finput)
     assert response.status_code == 302
-    ppt = api.get_participant(ppt_id, token=token)
+    ppt = api.get_participant(ppt_id, token=token_fixture)
     apt_ids = list(ppt.appointments.records.keys())
     last_apt_id = apt_ids[-1].split(":")[1]
     assert api.make_id(ppt_id, last_apt_id) in apt_ids
@@ -50,13 +50,13 @@ def test_apt_mod(client, apt_finput_mod):
     assert response.status_code == 200
 
 
-def test_apt_mod_post(client, apt_finput_mod, token):
+def test_apt_mod_post(client, apt_finput_mod, token_fixture):
     """Test apt_all endpoint."""
     apt_id = apt_finput_mod["inputAptId"]
-    apt = api.get_appointment(apt_id, token=token)
+    apt = api.get_appointment(apt_id, token=token_fixture)
 
     url = f"/appointments/{apt_id}/appointment_modify"
     response = client.post(url, data=apt_finput_mod)
     assert response.status_code == 302
-    new_apt = api.get_appointment(apt_id, token=token)
+    new_apt = api.get_appointment(apt_id, token=token_fixture)
     assert new_apt.data != apt.data
