+++ tests/functional/test_participants.py
@@ -1,6 +1,6 @@
 """Test participants endpoints."""
 
-from tests import utils as tutils
+from tests.conftest import participant_exists
 from babylab.src import api
 
 
@@ -22,12 +22,12 @@ def test_ppt_new(client):
     assert response.status_code == 200
 
 
-def test_ppt_new_post(client, ppt_finput, token):
+def test_ppt_new_post(client, ppt_finput, token_fixture):
     """Test ppt_all endpoint."""
-    ppt_id = api.get_next_id(token=token)
-    assert not tutils.participant_exists(ppt_id)
+    ppt_id = api.get_next_id(token=token_fixture)
+    assert not participant_exists(ppt_id)
     response = client.post("/participant_new", data=ppt_finput)
-    assert tutils.participant_exists(ppt_id)
+    assert participant_exists(ppt_id)
     assert response.status_code == 302
 
 
@@ -38,11 +38,11 @@ def test_ppt_mod(client, ppt_finput_mod):
     assert response.status_code == 200
 
 
-def test_ppt_mod_post(client, ppt_finput_mod, token):
+def test_ppt_mod_post(client, ppt_finput_mod, token_fixture):
     """Test ppt_all endpoint."""
-    ppt = api.get_participant(ppt_finput_mod["record_id"], token=token)
+    ppt = api.get_participant(ppt_finput_mod["record_id"], token=token_fixture)
     url = f"/participants/{ppt_finput_mod['record_id']}/participant_modify"
     response = client.post(url, data=ppt_finput_mod)
     assert response.status_code == 302
-    new_ppt = api.get_participant(ppt_finput_mod["record_id"], token=token)
+    new_ppt = api.get_participant(ppt_finput_mod["record_id"], token=token_fixture)
     assert new_ppt.data != ppt.data
