+++ tests/functional/test_questionnaires.py
@@ -24,10 +24,10 @@ def test_que_new(client, que_finput):
     assert response.status_code == 200
 
 
-def test_que_new_post(client, que_finput, token):
+def test_que_new_post(client, que_finput, token_fixture):
     """Test que_new endpoint."""
     ppt_id = que_finput["inputId"]
-    ppt = api.get_participant(ppt_id, token=token)
+    ppt = api.get_participant(ppt_id, token=token_fixture)
     que_ids = list(ppt.questionnaires.records.keys())
     last_que_id = que_ids[-1].split(":")[1]
     next_que_id = str(int(last_que_id) + 1)
@@ -35,7 +35,7 @@ def test_que_new_post(client, que_finput, token):
     url = f"/questionnaires/questionnaire_new?ppt_id={ppt_id}"
     response = client.post(url, data=que_finput)
     assert response.status_code == 302
-    ppt = api.get_participant(ppt_id, token=token)
+    ppt = api.get_participant(ppt_id, token=token_fixture)
     que_ids = list(ppt.questionnaires.records.keys())
     last_que_id = que_ids[-1].split(":")[1]
     assert api.make_id(ppt_id, last_que_id) in que_ids
@@ -48,14 +48,14 @@ def test_que_mod(client, que_finput_mod):
     assert response.status_code == 200
 
 
-def test_que_mod_post(client, que_finput_mod, token):
+def test_que_mod_post(client, que_finput_mod, token_fixture):
     """Test que_mod endpoint."""
     que_id = que_finput_mod["inputQueId"]
-    que = api.get_questionnaire(que_id, token=token)
+    que = api.get_questionnaire(que_id, token=token_fixture)
 
     url = f"/questionnaires/{que_id}/questionnaire_modify"
     response = client.post(url, data=que_finput_mod)
     assert response.status_code == 302
 
-    new_que = api.get_questionnaire(que_id, token=token)
+    new_que = api.get_questionnaire(que_id, token=token_fixture)
     assert new_que.data != que.data
